backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.0
azure-storage-blob>=12.19.0
twilio>=8.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
# face recognition (install separately if cmake/dlib available)
# face-recognition>=1.3.0
# opencv-python>=4.8.0
# numpy>=1.24.0
